requirements.txt
typer[all]
black
requests
python-dotenv
sqlalchemy
pymongo
motor==2.3.0
uuid
pydantic
validators
